Cleaned up the Hooks, moved everything where it belongs. Expanded on GatherComponentsEvent.java to be an abstract class.
